% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{enrichANAT}
\alias{enrichANAT}
\title{Method to enrich in anatomical terms provided by Bgee Database}
\usage{
enrichANAT(object, rid = 1, fData.tag = 1, sel.pval = c("P.Value",
  "adj.P.Val"), th.pval = 0.01, sel.feature = "genes", feature.null = "",
  bgee.data = c("rna_seq", "affymetrix", "est", "in_situ"),
  statistic = "fisher", verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{object}{Object of class \code{ResultSet}.}

\item{rid}{(default \code{1}) If given \code{ResultSet} hat has more than 
one result, id to select the result where the enrichment will be performed.}

\item{fData.tag}{(default \code{1}) Identifier of the annotation to be
used to mach the probes of \code{ResultSet} with gene identifier. In other
words, index of the element in \code{fData} that matches the omic dataset 
to be used in the enrichment.}

\item{sel.pval}{(default \code{"adj.P.Val"}) Name of the P.Value columns to 
filter datset's probes given \code{th.pval}.}

\item{th.pval}{(default \code{0.01}) Threshold used to include a probe (CpG)
into the enrichment analysis.}

\item{sel.feature}{(default \code{"genes"}) Name containing gene identifier
in \code{ResultSet}'s \code{fData}.}

\item{feature.null}{(default \code{""}) String identifier for non 
gene-assigned probes.}

\item{bgee.data}{(default \code{"affymetrix"}) Type of data requested to 
Bgee to perform the functinal enrichment. Can take any of them or more than
one of the possibles options.}

\item{statistic}{(default \code{"fisher"}) Statistic used to test the
enrichment in Bgee anatomical categories throught \code{topGO} R package.
Check \code{whichTests()} for available tests.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} messages
indicating the steps done by the method are raised.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} important
notes raised by the method are hide.}
}
\description{
TODO TODO TODO TODO TODO TODO TODO TODO TODO TODO TODO TODO TODO TODO TODO
}
\note{
Komljenovic A, Roux J, Robinson-Rechavi M, Bastian F;
BgeeDB, an R package for retrieval of curated expression datasets and for 
gene list enrichment tests. 
F1000Research (2016).

Bastian F, Parmentier G, Roux J, Moretti S, Laudet V, Robinson-Rechavi M;
Bgee: Integrating and Comparing Heterogeneous Transcriptome Data Among 
Species; Data Integration Life Sci. Lecture Notes in Computer Science (2008)
}
