% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{enrichGWAS}
\alias{enrichGWAS}
\title{Function to enrich GWAS results on GWAS Catalog}
\usage{
enrichGWAS(object, version = c("hg19", "hg38", "current"), verbose = FALSE,
  warnings = TRUE)
}
\arguments{
\item{object}{Object of class \code{ResultSet}.}

\item{version}{Version of the GWAS Catalog to be used. This argument can
take \code{"hg19"}, \code{"hg38"} and \code{"current"}. If it is set to 
\code{"current"} an update snapshot from GWAS Catalog will be downloaded,
otherwise stored version in \code{gwascat} will be used.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} messages
indicating the steps done by the method are raised.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} important
notes raised by the method are hide.}
}
\description{
This method takes a \code{ResultSet} and performs a query to GWAS Catalog
data giving a set of SNPs and retrieving the associated results from the
database.
}
\note{
Welter D, MacArthur J, Morales J, Burdett T, Hall P, Junkins H, Klemm A, 
Flicek P, Manolio T, Hindorff L, Parkinson H;
The NHGRI GWAS Catalog, a curated resource of SNP-trait associations.
Nucleic Acids Research (2014)
}
\seealso{
See \link{makeCurrentGwascat} for more information on updated
version of GWAS Catalog.
}

