% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{enrichPATHf}
\alias{enrichPATHf}
\title{Function to enrich results from differential studies on KEGG and GO 
databases}
\usage{
enrichPATHf(object, family, fData.exp = 2, fData.omic = 1,
  version = c("hg18", "hg19"), database = c("KEGG", "GO:CC", "GO:BP",
  "GO:MF"), sel.pval = c("P.Value", "adj.P.Val"), th.pval = 0.01,
  sel.feature = "genes", feature.id = "geneSymbol", feature.null = "",
  plot.fit = FALSE, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{object}{Object of class \code{ResultSet}.}

\item{family}{Family of exposures used to perform the enrichment analysis.}

\item{fData.exp}{(default \code{2}) Identifier of the exposures decrition 
to be used to mach the exposures of \code{ResultSet} with their families.}

\item{fData.omic}{(default \code{1}) Identifier of the annotation to be
used to mach the probes of \code{ResultSet} with gene identifier. In other
words, index of the element in \code{fData} that matches the omic dataset 
to be used in the enrichment.}

\item{version}{(default \code{"hg18"}) Can takes \code{"hg18"} and 
\code{"hg19"} and corresponds to the genome version to use for enrichment 
in KEGG/GO.}

\item{database}{(default \code{"KEGG"}) Can takes values \code{"KEGG"}, 
\code{"GO:CC"}, \code{"GO:BP"} and \code{"GO:MF"}; corresponding to 
the available enrichments.}

\item{sel.pval}{(default \code{"adj.P.Val"}) Name of the P.Value columns to 
filter datset's probes given \code{th.pval}.}

\item{th.pval}{(default \code{0.01}) Threshold used to include a probe (CpG)
into the enrichment analysis.}

\item{sel.feature}{(default \code{"genes"}) Name containing gene identifier
in \code{ResultSet}'s \code{fData}.}

\item{feature.id}{(default \code{"geneSymbol"}) Name of the format of the 
gene identifier. Run \link{supportedGeneIDs} from \code{geneLenDataBase} for 
more information. This must match with the identifier in \code{ResultSet}'s 
\code{fData}.}

\item{feature.null}{(default \code{""}) String identifier for non 
gene-assigned probes.}

\item{plot.fit}{(default \code{FALSE}) If set to \code{TRUE} calls plot 
\link{PWF} with default values.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} messages
indicating the steps done by the method are raised.}

\item{warnings}{(default \code{TRUE}) If set to \code{FALSE} important
notes raised by the method are hide.}
}
\description{
This method takes a \code{ResultSet} and performs an enrichment on KEGG
pathways or GO terms in base to a given set of genes. Argument 
\code{database} allows to select the database where the query will be 
performed. \code{sel.pval} allows to select the column of the resulting
table in \code{ResultSet} with the p-value to be used in the process.
Argument \code{th.pval} is used to exclude features using \code{sel.pval}
column. Argument \code{sel.feature} indicates the column that has the
gene associated to the feature (row in the table of results).
\code{feature.id} indicates the coding system for the \code{sel.feature}
column.
}
\note{
The Gene Ontology Consortium; Gene Ontology Consortium: going forward;
Nucl Acids Res (2015).

Ashburner et al.; Gene ontology: tool for the unification of biology;
Nat Genet (2000).

Kanehisa M, Furumichi M, Tanabe M, Sato Y, Morishima K; 
KEGG: new perspectives on genomes, pathways, diseases and drugs. 
Nucleic Acids Res (2017).

Kanehisa M, Sato Y, Kawashima M, Furumichi M, Tanabe M; 
KEGG as a reference resource for gene and protein annotation. 
Nucleic Acids Res (2016).

Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes and Genomes. 
Nucleic Acids Res. 28, 27-30 (2000).
}

