% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/EnirchResult-new.R,
%   R/join-ResultSet.R
\docType{class}
\name{EnrichResult-class}
\alias{EnrichResult-class}
\alias{EnrichSet}
\alias{join}
\title{Class \code{EnrichResult}}
\usage{
EnrichSet(fun_originA, fun_originB, search, database, eresult)

join(...)
}
\arguments{
\item{fun_originA}{Function used to create the \code{ResultSet} (aka. 
\code{enirchDIS}, ...).}

\item{fun_originB}{(optional) Function used to create the base object
of the new \code{ResultSet}. Usually \code{rexposome}'s functions (aka.
\code{assocGE}, ...).}

\item{search}{\code{character} used in the database query.}

\item{database}{\code{character} indicating the databased used.}

\item{eresult}{\code{list} with a single element, the result obtained from 
the query.}

\item{...}{Two or more object of class \code{EnrichResult}}
}
\description{
Class \code{EnrichResult}

Function to create a \code{ResultSet} from scratch

This function allows to combine multiple \code{\link{ResultSet}}
objects in a single one.
}
\section{Functions}{
\itemize{
\item \code{EnrichSet}: 

\item \code{join}: 
}}

\section{Slots}{

\describe{
\item{\code{fun_origin}}{\code{list} containing the methods used to create
the \code{EnrichResult} object/s.}

\item{\code{search}}{\code{list} containing the features used to query the
web databses and to create \code{EnrichResult} object/s.}

\item{\code{database}}{\code{list} containing the database queried to create
\code{EnrichResult} object/s.}

\item{\code{eresult}}{\code{list} containing the results obtained on quering
web databases}
}}

\examples{
data(air)
no2 <- enrichDIS(air, rid="NO2", sel.pval="P.Value", warnings=FALSE,
                 sel.feature="gene_assignment", translate.from="REFSEQ")
ben <- enrichDIS(air, rid="Ben", sel.pval="P.Value", warnings=FALSE,
                 sel.feature="gene_assignment", translate.from="REFSEQ")
join(no2, ben)
}
