% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postRexposome.R
\docType{package}
\name{postRexposome}
\alias{postRexposome}
\alias{postRexposome-package}
\title{postRexposome: Package for enrichment analysis in KEGG, GO, DisGeNET, GWAS Catalog and Begee for exposome-omic data analyis results generated with omicRexposome package.}
\description{
postRexposome: Package for enrichment analysis in KEGG, GO, DisGeNET, GWAS Catalog and Begee for exposome-omic data analyis results generated with omicRexposome package.
}
\section{KEGG/GO Enrichment}{

...
}

\section{DisGeNET Enrichment}{

...
}

\section{Bgee Enrichment}{

...
}

\section{GWAS Catalog Enrichment}{

...
}

\section{Plotting Enrichment}{

...
}

