% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/EnirchResult-cbind.R, R/EnirchResult-new.R
\docType{class}
\name{EnrichResult-class}
\alias{EnrichResult-class}
\alias{EnrichSet}
\alias{join}
\title{Class \code{EnrichResult}}
\usage{
join(...)

EnrichSet(fun_originA, fun_originB, search, database, eresult)
}
\arguments{
\item{...}{Two or more object of class \code{EnrichResult}}

\item{fun_originA}{Function used to create the \code{ResultSet} (aka. 
\code{enirchDIS}, ...).}

\item{fun_originB}{(optional) Function used to create the base object
of the new \code{ResultSet}. Usually \code{rexposome}'s functions (aka.
\code{assocGE}, ...).}

\item{search}{\code{character} used in the database query.}

\item{database}{\code{character} indicating the databased used.}

\item{eresult}{\code{list} with a single element, the result obtained from 
the query.}
}
\description{
Class \code{EnrichResult}

Function to create a \code{ResultSet} from scratch
}
\section{Functions}{
\itemize{
\item \code{join}: 

\item \code{EnrichSet}: 
}}
\section{Slots}{

\describe{
\item{\code{fun_origin}}{\code{list} containing the methods used to create
the \code{EnrichResult} object/s.}

\item{\code{search}}{\code{list} containing the elements used to query the
web databses and to create \code{EnrichResult} object/s.}

\item{\code{database}}{\code{list} containing the database queried to create
\code{EnrichResult} object/s.}

\item{\code{eresult}}{\code{list} containing the results obtained on quering
web databases}
}}

